+++ test/exfmt/comment_test.exs
@@ -151,6 +151,15 @@ defmodule Exfmt.CommentTest do
     test "single quote character" do
       assert extract_comments(~S(?\')) == {:ok, []}
     end
+
+    test "string in docstring" do
+      code = ~S(
+      """
+      debug "hello?"
+      """
+      )
+      assert extract_comments(code) == {:ok, []}
+    end
   end
 
   describe "merge/2" do