+++ test/exfmt/integration/sigil_test.exs
@@ -11,6 +11,9 @@ defmodule Exfmt.Integration.SigilTest do
     "~r[hello]" ~> "~r/hello/\n"
     "~r{hello}" ~> "~r/hello/\n"
     ~S"~r/\//" ~> "~r(/)\n"
+  end
+
+  test "r sigil 2" do
     ~S"~r/\/()/" ~> ~S"~r(/(\))" <> "\n"
   end
 
@@ -49,15 +52,7 @@ defmodule Exfmt.Integration.SigilTest do
     ~S(ø)
     """
   end
-  #
-  # TODO: Work out what is suppsed to happen here.
-  # I can't work out a way to render this in a fashion
-  # that makes the compiler happy.
-  # I'm wondering if there is a bug in the parser. There
-  # seems to be a bug in the Inspect protocol.
-  # https://github.com/elixir-lang/elixir/issues/6255
-  #
-  @tag :skip
+
   test "sigil containing new close char that will need to be escaped" do
     ~S"""
     ~R" \( \) / "