+++ lib/exfmt/context.ex
@@ -18,7 +18,7 @@ defmodule Exfmt.Context do
 
   @valid_layers ~W(list call keyword access negative sigil spec_lhs spec_rhs
                    tuple module_attribute map fn & do last_arg chain_call
-                   <<>>)a ++ Infix.infix_ops
+                   struct_name <<>>)a ++ Infix.infix_ops
 
   @doc """
   Push a new value onto the stack, signifying another layer in the code.