+++ CHANGELOG.md
@@ -11,8 +11,9 @@ Changelog
   - Calls to function name atoms from another function.
   - Module attributes with values that have a block.
   - Infix operators with module attribute assignment arguments.
-  - Struct patterns with pinned names.
+  - Struct patterns with pinned types.
   - Infix operators with captured fn arguments.
+  - Structs with unquoted types.
 - Fix: Parse comments from files containing `?\"`, `?\'`, or heredocs.
 - Fix: Avoid applying `do` syntactic sugar to unsupported block words.
 - Fix: Correctly escape sigils containing their close character.